ADVfixes

+++ more_functions_nested_loops/100-prime_factor.c
@@ -18,7 +18,7 @@ int main(void)
 	{
 		if (x % i == 0)
 		{
-			int j;
+			unsigned long int j;
 			int count = 0;
 
 			for (j = 2; j < i; j++)
